add query and insertion benchmark set up

+++ vesar/src/datasets/synthetic_data.rs
@@ -1,7 +1,7 @@
 use rand::rng;
 use rand::RngExt;
 
-pub fn generate_data(n: usize, dim: usize) -> Vec<Vec<f32>> { 
+pub fn generate_data(n: u64, dim: u64) -> Vec<Vec<f32>> { 
     let points: Vec<Vec<f32>> = (0..n).map(|_|
         (0..dim).map(
             |_| rng().random_range(0.0..1.0)
@@ -11,7 +11,7 @@ pub fn generate_data(n: usize, dim: usize) -> Vec<Vec<f32>> {
     return points;
 }
 
-pub fn generate_query(n: usize, dim: usize) -> Vec<Vec<f32>> {
+pub fn generate_query(n: u64, dim: u64) -> Vec<Vec<f32>> {
     let points: Vec<Vec<f32>> = (0..n).map(|_|
         (0..dim).map(
             |_| rng().random_range(0.0..1.0)
